refactor: study, studies api 병합
