+++ src/api/studies.ts
@@ -151,3 +151,42 @@ export const patchNotice = async (studyId: number, notice: string) => {
     });
     return res.data.message;
 };
+
+/* 스터디 생성 */
+export const createStudy = async ({
+    name,
+    category,
+    maxMembers,
+    region,
+    place,
+    schedules,
+    startTime,
+    endTime,
+    startDate,
+    endDate,
+    description,
+    externalLink,
+    studyType,
+    goals,
+    online,
+}: CreateStudy) => {
+    const response = await axiosInstance.post("studies", {
+        name,
+        category,
+        maxMembers,
+        region,
+        place,
+        schedules,
+        startTime,
+        endTime,
+        startDate,
+        endDate,
+        description,
+        externalLink,
+        studyType,
+        goals,
+        online,
+    });
+
+    return response.data;
+};
