--- src/api/study.ts
@@ -1,40 +1,0 @@
-import { axiosInstance } from "./index";
-
-/* 스터디 생성 */
-export const createStudy = async ({
-    name,
-    category,
-    maxMembers,
-    region,
-    place,
-    schedules,
-    startTime,
-    endTime,
-    startDate,
-    endDate,
-    description,
-    externalLink,
-    studyType,
-    goals,
-    online,
-}: CreateStudy) => {
-    const response = await axiosInstance.post("studies", {
-        name,
-        category,
-        maxMembers,
-        region,
-        place,
-        schedules,
-        startTime,
-        endTime,
-        startDate,
-        endDate,
-        description,
-        externalLink,
-        studyType,
-        goals,
-        online,
-    });
-
-    return response.data;
-};
